resume-matchmaker/resume-matchmaker-backend/requirements.txt
blinker==1.7.0
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
Flask==3.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pycparser==2.22
pymongo==4.6.3
pypdfium2==4.28.0
python-docx==1.1.0
typing_extensions==4.11.0
Werkzeug==3.0.2
